+++ packages/authhero/src/routes/universal-login/common.tsx
@@ -116,7 +116,7 @@ export async function initJSXRoute(
     },
     client,
     tenant,
-    session: loginSession,
+    loginSession,
   };
 }