+++ packages/drizzle/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @authhero/drizzle
 
+## 0.2.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.51.0
+
 ## 0.2.10
 
 ### Patch Changes